Add files via upload
(CC) Whale Talk was a very interesting project to build, but if Dory from Finding Nemo can do it, then so can a computer. It involves Vectors and iterating through them in order to find values such as vowels. The basic rules of whale talk are that there are no consonants. The vowels 'u' and 'e' also get doubled since they are extra-long in whale talk.